requirements.txt
# Core dependencies
streamlit==1.29.0
python-dotenv==1.0.0
O365==2.0.31
cryptography==41.0.7

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
pytesseract==0.3.10
pdf2image==1.16.3

# NLP and text processing
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
pyyaml==6.0.1
pandas==2.1.4
numpy>=1.24.0,<2.0.0
click==8.1.7
colorama==0.4.6
openai>=1.50.0  # For enhanced OCR and PDF processing

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0 